pass a callback to return disabled flag
because riot continues to call update on child elements even if they're
currently not being mounted (and apparently react doesn't?), we need a
way to determine whether the waypoint should actually update to
eliminate things like the double fetching problem i was having with the
infinite-scroll example. previously, i was checking a value on the
waypoint's parent from the waypoint, which coupled things too closely
and interfered with the possibility of having multiple waypoints as
children on a single component.